clean up project tree and add DnD hook
