--- src/components/Canvas.scss
@@ -1,12 +1,0 @@
-@import "../styles/variables.scss";
-
-.canvas {
-  text-align: center;
-  margin: auto;
-  margin-left: $sidebar-width;
-  margin-top: 20px;
-  width: 600px;
-  height: 600px;
-  border: 2px solid red;
-  z-index: -1;
-}
