+++ src/components/Canvas/UrlImage.js
@@ -3,7 +3,6 @@ import { Image } from 'react-konva';
 import useImage from 'use-image';
 
 export default function UrlImage(props) {
-  console.log("urlImage props", props);
   const [img] = useImage(props.image.src);
   return (
     <Image
@@ -13,6 +12,7 @@ export default function UrlImage(props) {
       offsetX={img ? img.width / 2 : 0}
       offsetY={img ? img.height / 2 : 0}
       draggable
+      resizeEnabled
     />
   )
 };
