+++ src/components/Home/Splash.js
