+++ src/components/Left/Left.js
@@ -1,17 +1,20 @@
 import React from 'react'
-
-import Toolbar from './LeftDrawer/Toolbar/Toolbar'
 import './Left.scss'
 
+import Files from './Files'
+import Templates from './Templates'
+import Toolbar from './Toolbar'
+
 export default function Left(props) {
-  console.log("left props", props)
   return (
     <div className="left">
+      <Files/>
+      <Templates/>
       <Toolbar
+         imagesData={props.imagesData}
          onDragStart={props.onDragStart}
          onDrop={props.onDrop}
-         onDragOver={props.onDragOver}
-         imgData={props.imgData}
+        onDragOver={props.onDragOver}
       />
     </div>
   )
