+++ src/components/Left/Left.scss
@@ -1,4 +1,4 @@
-@import "../styles/variables.scss";
+@import "../../styles/variables.scss";
 
 .left {
   position: fixed;
