+++ src/components/Right/Right.js
